create a hook called useSearchResults
