+++ src/hooks/useSearchResults.type.ts
@@ -1,0 +1,7 @@
+export type UseSearchResults = {
+    searchQuery: string;
+    currentPage: number;
+    shouldFetchSearchResults: boolean;
+    setShowNoResultsMessage: (value: boolean) => void;
+    setSearchError: (value: boolean) => void;
+}
